Simplify the data loading for conversation data
