+++ services/agora/src/composables/useConversationData.ts
@@ -1,14 +1,18 @@
-import { storeToRefs } from "pinia";
-import { useLoginIntentionStore } from "src/stores/loginIntention";
-import { useConversationStore } from "src/stores/conversation";
-import { onMounted } from "vue";
+import { ref, computed, onMounted, watch } from "vue";
 import { useRoute } from "vue-router";
+import { storeToRefs } from "pinia";
+import type { ExtendedConversation } from "src/shared/types/zod";
+import { useHomeFeedStore } from "src/stores/homeFeed";
+import { useBackendPostApi } from "src/utils/api/post";
+import { useAuthenticationStore } from "src/stores/authentication";
+import { useLoginIntentionStore } from "src/stores/loginIntention";
 
 export function useConversationData() {
-  const conversationStore = useConversationStore();
-  const { conversationData, conversationLoaded } =
-    storeToRefs(conversationStore);
-
+  const { fetchPostBySlugId } = useBackendPostApi();
+  const { emptyPost } = useHomeFeedStore();
+  const { isGuestOrLoggedIn, isAuthInitialized } = storeToRefs(
+    useAuthenticationStore()
+  );
   const route = useRoute();
 
   const {
@@ -17,14 +21,18 @@ export function useConversationData() {
     clearReportUserContentIntention,
   } = useLoginIntentionStore();
 
+  // Simple reactive state
+  const conversationData = ref<ExtendedConversation>(emptyPost);
+  const isLoading = ref(false);
+
+  const hasConversationData = computed(
+    () => conversationData.value.metadata.conversationSlugId !== ""
+  );
+
   // Clear intentions on initialization
   clearVotingIntention();
   clearOpinionAgreementIntention();
   clearReportUserContentIntention();
-
-  onMounted(async () => {
-    await loadConversationData();
-  });
 
   function getConversationSlugId(): string | null {
     const isConversationRoute =
@@ -38,33 +46,82 @@ export function useConversationData() {
       return null;
     }
 
-    return Array.isArray(route.params.postSlugId)
-      ? route.params.postSlugId[0]
-      : route.params.postSlugId;
+    const postSlugId = (route.params as { postSlugId?: string | string[] })
+      .postSlugId;
+    return Array.isArray(postSlugId) ? postSlugId[0] : postSlugId || null;
   }
 
-  async function loadConversationData(refresh = false) {
+  async function loadConversationData() {
     const slugId = getConversationSlugId();
+    if (!slugId) return false;
 
-    if (slugId) {
-      return await conversationStore.loadConversationData(slugId, refresh);
+    // Wait for auth to be initialized
+    if (!isAuthInitialized.value) {
+      await new Promise<void>((resolve) => {
+        const unwatch = watch(isAuthInitialized, (isReady) => {
+          if (isReady) {
+            unwatch();
+            resolve();
+          }
+        });
+      });
     }
 
-    return false;
+    // Clear old data and start loading
+    conversationData.value = emptyPost;
+    isLoading.value = true;
+
+    try {
+      const response = await fetchPostBySlugId(slugId, isGuestOrLoggedIn.value);
+      if (response != null) {
+        conversationData.value = response;
+        return true;
+      } else {
+        conversationData.value = emptyPost;
+        return false;
+      }
+    } catch (error) {
+      console.error("Failed to load conversation data:", error);
+      conversationData.value = emptyPost;
+      return false;
+    } finally {
+      isLoading.value = false;
+    }
   }
 
   function refreshConversation(done: () => void) {
     setTimeout(() => {
       void (async () => {
-        await loadConversationData(true);
+        await loadConversationData();
         done();
       })();
     }, 500);
   }
 
+  // Watch for route parameter changes to load new conversation data
+  watch(
+    () => (route.params as { postSlugId?: string | string[] }).postSlugId,
+    async (newSlugId, oldSlugId) => {
+      if (newSlugId && newSlugId !== oldSlugId) {
+        await loadConversationData();
+      }
+    },
+    { immediate: false }
+  );
+
+  onMounted(async () => {
+    await loadConversationData();
+  });
+
   return {
+    // State
     conversationData,
-    conversationLoaded,
+    isLoading,
+
+    // Computed
+    hasConversationData,
+
+    // Actions
     loadConversationData,
     refreshConversation,
   };
